Installed - Laravel Passport
What
=================
Installed: Laravel Passport, a package that handles API authentication.
I've also registered and installed the migrations to fit our proposed
database, as well as the .vue components for the frontend. So in the
future when we get to the admin menu, use these components as a starting
point!

Issue ID: #4