Moved - Public files to resources to avoid anything being in "build"
What
=================
Moved: Public files that were in public folder before, even those that
were placed there in the laravel installation, into the resources
folder. Webpack.mx.js will copy the necessaries that sit directly in the
resources folder.

Why
=================
So that if there's a conflict or something fucky wucky, we can just
delete the whole folder.

Issue ID: N/A